add friends list

+++ client/src/app/user/post/post.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { Post, Like } from 'src/app/shared/models/post.model';
+import { Component, OnInit, Input, Output, EventEmitter, ViewChild, ElementRef } from '@angular/core';
+import { Post, Like, Comment } from 'src/app/shared/models/post.model';
 import { ServerService } from 'src/app/shared/services/http/http.service';
 
 @Component({
@@ -9,11 +9,16 @@ import { ServerService } from 'src/app/shared/services/http/http.service';
 })
 export class PostComponent implements OnInit {
   @Input() post: Post;
+  @Input() canDelete: boolean;
+  @Input() userId: number;
   @Output() deletePost = new EventEmitter<Post>();
+  @ViewChild('commentInput', { static: false }) commentInput: ElementRef;
 
   constructor(private serverService: ServerService) { }
 
   ngOnInit() {
+    console.log(this.post);
+    
   }
 
   public isLiked() {
@@ -24,27 +29,66 @@ export class PostComponent implements OnInit {
     return this.post.userLike === Like.dislike;
   }
 
+  public getComments() {
+    return this.post.comments;
+  }
+
+  public onCreateComment() {
+    const comment = (<HTMLInputElement>this.commentInput.nativeElement).value;
+    if (!comment) {
+      return;
+    }
+
+    this.serverService.createComment({ text: comment, postId: this.post.id }).subscribe((result: any) => {
+      this.post.comments.push({
+        firstname: result.firstname,
+        lastname: result.lastname,
+        id: result.id,
+        text: result.text,
+        userId: result.userId
+      });
+    });
+  }
+
+  public onDeleteComment(comment: Comment) {
+    if (!confirm('Delete this comment?')) {
+      return;
+    }
+
+    this.serverService.deleteComment(comment.id).subscribe((result) => {
+      this.post.comments.splice(this.post.comments.indexOf(comment), 1);
+    });
+  }
+
   public onLike() {
     if (this.isLiked()) {
       this.serverService.deleteLike(this.post.userLikeId).subscribe((result) => {
         console.log(result);
         this.post.userLikeId = null;
+        this.post.userLike = null;
+        this.post.likes--;
       })
     } else {
       if (this.post.userLike) {
         this.serverService.updateLike({
           likeId: this.post.userLikeId,
-          type: Like.dislike
-        }).subscribe((result: any) => {
-          this.post.userLike = result.type;
+          type: Like.like
+        }).subscribe((like: any) => {
+          this.post.userLike = like.type;
+          this.post.userLikeId = like.id;
+          this.post.likes++;
+          this.post.dislikes--;
         });
       } else {
         this.serverService.likePost({ type: Like.like, postId: this.post.id }).subscribe((like: any) => {
           console.log(like);
+          this.post.userLike = like.type;
           this.post.userLikeId = like.id;
+          this.post.likes++;
         });
       }
     }
+    console.log(this.post);
   }
 
   public onDislike() {
@@ -52,25 +96,37 @@ export class PostComponent implements OnInit {
       this.serverService.deleteLike(this.post.userLikeId).subscribe((result) => {
         console.log(result);
         this.post.userLikeId = null;
+        this.post.userLike = null;
+        this.post.dislikes--;
       })
     } else {
       if (this.post.userLike) {
         this.serverService.updateLike({
           likeId: this.post.userLikeId,
-          type: Like.like
-        }).subscribe((result: any) => {
-          this.post.userLike = result.type;
+          type: Like.dislike
+        }).subscribe((like: any) => {
+          this.post.userLike = like.type;
+          this.post.userLikeId = like.id;
+          this.post.likes--;
+          this.post.dislikes++;
         });
       } else {
         this.serverService.likePost({ type: Like.dislike, postId: this.post.id }).subscribe((like: any) => {
           console.log(like);
+          this.post.userLike = like.type;
           this.post.userLikeId = like.id;
+          this.post.dislikes++;
         });
       }
     }
+    console.log(this.post);
   }
 
   public onDeletePost() {
+    if (!confirm('Delete this post?')) {
+      return;
+    }
+
     this.serverService.deletePost(this.post.id).subscribe((result) => {
       console.log(result);
       this.deletePost.emit(this.post);
